test(cli): receive an accepted note deletion on pull
Fast-forward a second checkout at the deletion parent so the removed file is gone, remaining bytes match, and ancestry is kept.
